+++ src/main.js
@@ -3,17 +3,17 @@ import App from './App.vue';
 import { createRouter, createWebHistory } from 'vue-router';
 
 import Dashboard from './components/Dashboard.vue';
-import Fazendas from './components/Fazendas.vue';
+import UnitList from './components/UnitList.vue';
 import Animais from './components/AnimalList.vue';
-import Vacinas from './components/Vacinas.vue';
+import VaccineList from './components/VaccineList.vue';
 import Calendario from './components/Calendario.vue';
 import Notificacoes from './components/Notificacoes.vue';
 
 const routes = [
   { path: '/dashboard', component: Dashboard },
-  { path: '/units', component: Fazendas },
+  { path: '/units', component: UnitList },
   { path: '/animals', component: Animais },
-  { path: '/vaccines', component: Vacinas },
+  { path: '/vaccines', component: VaccineList },
   { path: '/calendar', component: Calendario },
   { path: '/notifications', component: Notificacoes },
 ];
